+++ src/tr2/game/phase/phase_stats.h
@@ -6,8 +6,8 @@
 
 typedef struct {
     bool show_final_stats;
-    int32_t fade_in_time;
-    int32_t fade_out_time;
+    double fade_in_time;
+    double fade_out_time;
 } PHASE_STATS_ARGS;
 
 PHASE *Phase_Stats_Create(PHASE_STATS_ARGS args);
